Renamed App to project-details-page, created separate folders for i18n, common logic separated from page, ProjectType contains types of owner, coordinates, networks, project-details moved to  project-page folder
