+++ src/common/types/projects/coordinates.type.ts
@@ -1,0 +1,4 @@
+export type CoordinatesType = {
+    latitude: number,
+    longitude: number,
+};
