+++ src/common/types/projects/networks-links.type.ts
@@ -1,0 +1,5 @@
+export type NetworksLinks = {
+    FACEBOOK: string,
+    INSTAGRAM: string,
+    YOUTUBE: string
+};
