+++ src/common/types/projects/owner.type.ts
@@ -1,0 +1,5 @@
+export type OwnerType = {
+    ownerId: string,
+    firstName: string,
+    lastName: string,
+};
