+++ src/common/types/projects/projects.ts
@@ -1,1 +1,1 @@
-export { type ProjectsType } from './projects.type';
+export { type ProjectType } from './project.type';
