--- src/common/types/projects/projects.type.ts
@@ -1,17 +1,0 @@
-export type ProjectsType = {
-  id: string,
-  title: string,
-  summary: string,
-  owner: string
-  description: string,
-  logo: string,
-  geoCoordinates: string,
-  creationDate: string,
-  status: string,
-  ownerSum: number,
-  collectedSum: number,
-  startSum: number,
-  collectDeadline: string,
-  goalSum: number,
-  goalDeadline: string,
-};
