+++ src/common/types/types.ts
@@ -1,4 +1,5 @@
 export { type AppDispatch, type AsyncThunkConfig, type RootState } from './app/app';
 export { type HttpErrorDto, type HttpOptions } from './http/http';
 export { type NotificationPayload } from './notification/notification';
-export { type ProjectsType } from './projects/projects';
+export { type ProjectType } from './projects/projects';
+
