+++ src/components/common/common.ts
@@ -1,1 +1,2 @@
+export { CustomTimeline } from './time-line/time-line';
 export { Toast } from './toast/toast';
