+++ src/i18n/en-translation.ts
@@ -1,0 +1,29 @@
+const enTranslation = {
+  projects: 'Projects',
+  about_us: 'About us',
+  log_in: 'Log in',
+  sign_in: 'Sign in',
+  search_field: 'Search field',
+  share_button: 'Share',
+  learn_more_button: 'Learn more',
+  footer_sign: 'All rights reserved',
+  statuses: {
+    NEW: 'New',
+    GATHERING_FOR_START: 'Gathering for start',
+    EARNING: 'Earning',
+    SUSPENDED: 'Suspended',
+    FAILED: 'Failed',
+    SUCCESSFUL: 'Successful',
+  },
+  buttons: {
+    subscribe: 'Subscribe',
+    donate: 'Donate',
+    support: 'Support',
+    location: 'Location',
+  },
+  owner: 'Owner',
+  description: 'Description',
+};
+
+export { enTranslation };
+
