+++ src/i18n/uk-translation.ts
@@ -1,0 +1,28 @@
+const ukTranslation = {
+  projects: 'Проекти',
+  about_us: 'Про нас',
+  log_in: 'Увійти',
+  sign_in: 'Приєднатись',
+  search_field: 'Пошук',
+  share_button: 'Поділитися',
+  learn_more_button: 'Дізнатися більше',
+  footer_sign: 'Всі права захищені',
+  statuses: {
+    NEW: 'Новий',
+    GATHERING_FOR_START: 'Збирає на старт',
+    EARNING: 'Заробляє',
+    SUSPENDED: 'Призупинений',
+    FAILED: 'Провалений',
+    SUCCESSFUL: 'Успішний',
+  },
+  buttons: {
+    subscribe: 'Підписатися',
+    donate: 'Зробити внесок',
+    support: 'Підтримати',
+    location: 'Місцезнаходження',
+  },
+  owner: 'Власник',
+  description: 'Опис',
+};
+
+export { ukTranslation };
