+++ src/index.tsx
@@ -1,4 +1,4 @@
-import './services/i18n';
+import './i18n/i18n';
 
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
@@ -6,11 +6,11 @@ import { Provider } from 'react-redux';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { store } from 'store/store';
 
-import { App } from './components/app/app';
 import { ProjectsPage } from './components/app/projects-page';
 import { Toast } from './components/common/common';
 import { Footer } from './components/footer/footer';
 import { Header } from './components/header/header';
+import { ProjectDetailsPage } from './components/project-page/project-details-page';
 
 const root = createRoot(document.getElementById('root') as HTMLElement);
 
@@ -21,7 +21,7 @@ root.render(
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<ProjectsPage/>}/>
-          <Route path="projects/1" element={<App/>}/>
+          <Route path="projects/1" element={<ProjectDetailsPage/>}/>
         </Routes>
       </BrowserRouter>
       <Footer/>
