+++ src/utils/function-generate-google-maps-link.tsx
@@ -1,0 +1,8 @@
+import { ProjectType } from '../common/types/projects/project.type';
+
+export function generateGoogleMapsLink(project: ProjectType): string {
+  const { coordinates } = project;
+  const { latitude, longitude } = coordinates;
+
+  return `https://www.google.com/maps/place/${latitude},${longitude}`;
+}
